move all db code to db folder
