+++ backend/src/db/song.js
@@ -1,0 +1,8 @@
+import sql from './db.js'
+
+export async function getSong(id) {
+    return await sql`
+        SELECT * 
+        FROM    songs s
+        WHERE   s.id = ${id}`
+}
